third_party/WebKit/LayoutTests/overflow/overflow-basic-002.html

<!doctype html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<style>
#container {
  overflow: auto;
  border: 1px solid black;
  width: 200px;
  height: 100px;
  padding: 50px;
}
#target {
  width: 300px;
  height: 200px;
  background-color: rgba(0, 255, 0, 0.3);
}
</style>
<p>Overflow with container padding: container block_end, inline_end padding, are not part of overflow.</p>
<div id="container">
  <div id="target"></div>
</div>
<script>

var container = document.querySelector('#container');
var target = document.querySelector('#target');

test(function() {
  var containerStyle = window.getComputedStyle(container);
  assert_equals(container.scrollHeight, target.offsetHeight +
    parseInt(containerStyle.paddingTop));
  assert_equals(container.scrollWidth, target.offsetWidth +
    parseInt(containerStyle.paddingTop));
}, "overflow with padding");

test(function() {
  target.style.height = "400px";
  var containerStyle = window.getComputedStyle(container);
  assert_equals(container.scrollHeight, target.offsetHeight +
    parseInt(containerStyle.paddingTop));
  assert_equals(container.scrollWidth, target.offsetWidth +
    parseInt(containerStyle.paddingTop));
}, "overflow with paddding, after resize");
</script>
